reorganized experience components
