+++ src/lib/formHelpers/formValidator.js
@@ -1,4 +1,4 @@
-import {monthFullNames} from 'components/Experiences/AddOrEditExperienceForm/staticData';
+import {monthFullNames} from 'lib/formHelpers/staticData.js';
 
 // Validate add new experience form (Work/Education/Accomplishment)
 const experienceValidator = values => {
